Added basic enemies
They follow their own set paths around. They also have the ability to
change and chase the player.
Also added a GameManager, which currently can reset the level when the
player is caught.
Right now, the enemies just need to be changed to target the player when
the player walks into light.
Also, they walk through walls. That's still an issue