Updated some enemy AI and brightness
Also added a brightness adjuster. Use R and F for that. Now the enemies
only chase the player if they can see the player. Also, they stop
chasing the player if they lose line of sight (after going to the
player's last known position) and resume guarding.
Enemies can sometimes get stuck on walls when this happens if their next
point is through a wall.