Added a camera that can detect the player
Right now, the indicator is just a sprite that goes through walls. It
does not detect the player through walls however. Camera turns blue when
it sees the player, but the player is not lit (undetected) and red when
the player is lit (detected). Then baddies will come over to say hi.
Detection is fairly accurate. Not perfect but that could be a sprite
issue. Also i readded the map since it was missing.
Sorry this is so late but I was working on another project tonight as
well